Ladder - step3 - refactor (도메인 잔여로직 수정)
